Prevent deleting records that are still in use

+++ modules/companies/delete.php
@@ -1,8 +1,15 @@
 <?php
 require '../../config/config.php';
+
 $id = (int)($_GET['id'] ?? 0);
 if ($id > 0) {
-    mysqli_query($conn, "DELETE FROM companies WHERE id = $id");
+    $usageError = '';
+    if (!itm_can_delete_record($conn, 'companies', 'id', $id, 0, $usageError)) {
+        $_SESSION['crud_error'] = $usageError;
+    } else {
+        mysqli_query($conn, "DELETE FROM companies WHERE id = $id");
+    }
 }
+
 header('Location: index.php');
 exit;
